backend/pyproject.toml
[project]
name = "doc-vault"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "langchain-postgres>=0.0.16",
    "pgvector>=0.3.6",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.7",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-keycloak>=7.0.2",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
